Fix types in NVRTC testing

+++ test/test_cos_pi_nvrtc_float.cpp
@@ -21,7 +21,7 @@
 typedef float float_type;
 
 const char* cuda_kernel = R"(
-typedef double float_type;
+typedef float float_type;
 #include <boost/math/special_functions/cos_pi.hpp>
 extern "C" __global__ 
 void test_cos_pi_kernel(const float_type *in1, const float_type*, float_type *out, int numElements)
